Implemented basic training for CNNs

+++ src/main/java/net/echo/brain4j/utils/Vector.java
@@ -76,7 +76,7 @@ public class Vector implements Cloneable, Iterable<Double> {
         data[index] += (float) value;
     }
 
-    public double get(int index) {
+    public float get(int index) {
         return data[index];
     }
